Fixing duplication errors in search filters tests and award filters
